commit with test

+++ com.cg.homeloan/src/main/java/com/cg/homeloan/repositories/ILoanAgreementRepository.java
@@ -1,11 +1,9 @@
 package com.cg.homeloan.repositories;
 
-import java.util.List;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.cg.homeloan.entities.LoanAgreement;
 
-public interface ILoanAgreementRepository extends JpaRepository<LoanAgreement, Long> {
+public interface ILoanAgreementRepository extends JpaRepository<LoanAgreement, Integer> {
 	
 //	public LoanAgreement addLoanAgreement(LoanAgreement loanAgreement);
 //	public LoanAgreement updateLoanAgreement(LoanAgreement loanAgreement);
